Disable webhooks during API docs generation
Since the webhook API endpoints for the demo DB do not actually exist, we should disable those calls to not wait unnecessary for responses that we will never get.
Additionally it is probably better to not send requests to example.com, which is the URL used in the demo webhooks.