Improve handling to test and submission statuses
The status of tests can now be set explicitly. If set explicitly, the test will always have the status which is the worst of the explicitly set status and the worst of the test result statuses (i.e. you can not set the status of a test to be better than the worst status of any of its test results).
For submissions, we store the status in the database explicitly now, but it cannot be set explicitly be the user. It is updated whenever the statuses of any of it's tests change.